added "omitOutputtingOriginalText: true" header line
From documentation
http://www.opencalais.com/wp-content/uploads/folder/Thomson-Reuters-Open-Calais-API-User-GuideR10_6.pdf
Page 20

Description: Use this parameter to exclude the submitted text from the output, thus reducing the size of the response.
Values: true,  false
Default Value: false (The original text is included in the output.)
Remarks: By default, Open Calais returns the submitted text. Set this parameter to true to exclude the original text from the output.
It is highly recommended to use this header for large input files